requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.23.2
supabase==1.0.3
requests==2.31.0
pdfplumber==0.10.3
nltk==3.8.1
python-dotenv==1.0.1